fix also now 90 lines

+++ server.js
@@ -2,17 +2,14 @@ const { Client } = require('discord.js');
 const yt = require('ytdl-core');
 const tokens = require('./tokens.json');
 const client = new Client();
-
 let queue = {};
-
 const commands = {
 	'play': (msg) => {
-		if (queue[msg.guild.id] === undefined) return msg.channel.send(':x: Add some songs to the queue first with ${tokens.prefix}add`);
+		if (queue[msg.guild.id] === undefined) return msg.channel.send(`:x: Add some songs to the queue first with ${tokens.prefix}add`);
 		if (!msg.guild.voiceConnection) return commands.join(msg).then(() => commands.play(msg));
 		if (queue[msg.guild.id].playing) return msg.channel.send(':x: Already Playing');
 		let dispatcher;
 		queue[msg.guild.id].playing = true;
-
 		console.log(queue);
 		(function play(song) {
 			console.log(song);
@@ -81,14 +78,10 @@ const commands = {
 		let tosend = ['```xl', tokens.prefix + 'join : "Join Voice channel of msg sender"',	tokens.prefix + 'add : "Add a valid youtube link to the queue"', tokens.prefix + 'queue : "Shows the current queue, up to 15 songs shown."', tokens.prefix + 'play : "Play the music queue if already joined to a voice channel"', '', 'the following commands only function while the play command is running:'.toUpperCase(), tokens.prefix + 'pause : "pauses the music"',	tokens.prefix + 'resume : "resumes the music"', tokens.prefix + 'skip : "skips the playing song"', tokens.prefix + 'time : "Shows the playtime of the song."',	'volume+(+++) : "increases volume by 2%/+"',	'volume-(---) : "decreases volume by 2%/-"',	'```'];
 		msg.channel.send(tosend.join('\n'));
 	},
-	'reboot': (msg) => {
-		if (msg.author.id == tokens.adminID) process.exit(); //Requires a node module like Forever to work.
-	}
+	'reboot': (msg) => { if (msg.author.id == tokens.adminID) process.exit(); } //Requires a node module like Forever to work.
 };
 
-client.on('ready', () => {
-	console.log('Ready!');
-});
+client.on('ready', () => { console.log('Ready!'); });
 
 client.on('message', msg => {
 	if (!msg.content.startsWith(tokens.prefix)) return;
